Moved to WebGL2 + Color range from 255 to 0 (instead of 1-0)

+++ src/renderer-helper.ts
@@ -77,7 +77,7 @@ export class RendererHelper {
         gl.bindBuffer(gl.ARRAY_BUFFER, colorBuffer);
         gl.bufferData(
             gl.ARRAY_BUFFER,
-            new Float32Array(BoxObject.colors),
+            new Uint8Array(BoxObject.colors),
             gl.STATIC_DRAW,
         );
